requirements.txt
python-dotenv
pandas
spotipy
lyricsgenius
langchain
langchain-groq
langgraph
pydantic
streamlit